requirements.txt
playwright
beautifulsoup4
lxml
pdfplumber
pandas
requests
python-dotenv
pydantic
tenacity
urllib3
tqdm
pint
pyyaml
